differ modules js
