+++ js/main.js
@@ -1,7 +1,16 @@
-import { renderCard, renderDataCards } from "./template.js";
-import "./form.js";
-import "./map.js";
-import { sendFormData } from "./api.js";
-import { createSuccessSubmit,renderErrorMessage } from "./messages.js";
+import { renderPopups } from "./map.js";
+import { setUserFormSubmit } from "./form.js";
+import { createSuccessSubmit,renderErrorMessage, createErrorMessageGet } from "./messages.js";
+import { getData } from "./api.js";
 
-sendFormData(createSuccessSubmit, renderErrorMessage);
+getData((data) => {
+  renderPopups(data.slice(0,10));
+/*  houseType.addEventListener("input", () => {
+    markerGroup.clearLayers();
+    const objNew = newArray.filter(isHouse);
+    renderPopups(objNew.slice(0,10));
+  });*/
+}, createErrorMessageGet);
+
+setUserFormSubmit(createSuccessSubmit, renderErrorMessage);
+
